requirements.txt
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
dvc[s3]==3.51.1
boto3==1.34.106
pyyaml==6.0.1
